Products page has a search by category and text feature now.

+++ MVCECommerceRemake/Models/Customers.cs
@@ -5,7 +5,7 @@ namespace MVCECommerceRemake.Models
 {
     public partial class Customers
     {
-        public double CustomerId { get; set; }
+        public int CustomerId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Phone { get; set; }
